refactor: split backend into modules and add comments for readability
Split backend/main.py (630 lines) into focused modules:
- modules/config.py: constants, env vars, logging, error codes
- modules/validation.py: file/UUID validation, filename sanitization
- modules/presets.py: conversion presets, custom param validation
- modules/converter.py: vtracer wrapper with error handling
- modules/progress.py: progress store, SSE streaming, cleanup

main.py is now a thin orchestration layer (~250 lines) with routes
and app setup.

Added comments to complex frontend logic:
- post.ts: JSDoc on API client, base64 encoding, SSE progress
- +page.svelte: comments on preview URL memoization, batch upload,
  directory drag-drop, ZIP download

Closes #135
